Fix create unitypackage (#70)
* Run postprocessor only if not a package export

* Rename meta files
Rename meta files otherwise unity will create an empty folder when package is created

* Mark the android-wrapper folder as Android Platform only

+++ CleverTap/Editor/AndroidPostImport.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Linq;
 using UnityEditor;
 
@@ -5,8 +6,14 @@ namespace CleverTapSDK.Private
 {
     class AndroidPostImport : AssetPostprocessor
     {
+        private static readonly string ctExportCommandLineArg = "-ct-export";
+
         static void OnPostprocessAllAssets(string[] importedAssets, string[] deletedAssets, string[] movedAssets, string[] movedFromAssetPaths, bool didDomainReload)
         {
+            string[] cmdArgs = Environment.GetCommandLineArgs();
+            if (cmdArgs.Contains(ctExportCommandLineArg))
+                return;
+
             if (AssetDatabase.IsValidFolder("Assets/CleverTap/Plugins/Android/clevertap-android-wrapper") && importedAssets?.Length > 0 && importedAssets.Contains("Assets/CleverTap/Plugins/Android/clevertap-android-wrapper"))
             {
                 AssetDatabase.DeleteAsset("Assets/CleverTap/Plugins/Android/clevertap-android-wrapper.androidlib");
